+++ src/Netstr/Extensions/MessagingExtensions.cs
@@ -1,9 +1,9 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Netstr.Messaging;
-using Netstr.Messaging.EventHandlers;
-using Netstr.Messaging.EventHandlers.Replaceable;
+using Netstr.Messaging.Events.Handlers;
+using Netstr.Messaging.Events.Handlers.Replaceable;
+using Netstr.Messaging.Events.Validators;
 using Netstr.Messaging.MessageHandlers;
-using Netstr.Messaging.MessageHandlers.Events;
 using Netstr.Messaging.WebSockets;
 using Netstr.Middleware;
