+++ src/Netstr/Messaging/Events/Validators/EventCreatedAtValidator.cs
@@ -2,7 +2,7 @@
 using Netstr.Messaging.Models;
 using Netstr.Options;
 
-namespace Netstr.Messaging.MessageHandlers.Events
+namespace Netstr.Messaging.Events.Validators
 {
     /// <summary>
     /// <see cref="IEventValidator"/> which validates event's created_at is not too far in the past or in the future.
