+++ src/Netstr/Messaging/Events/Validators/EventHashValidator.cs
@@ -3,7 +3,7 @@ using System.Security.Cryptography;
 using System.Text.Encodings.Web;
 using System.Text.Json;
 
-namespace Netstr.Messaging.MessageHandlers.Events
+namespace Netstr.Messaging.Events.Validators
 {
     /// <summary>
     /// <see cref="IEventValidator"/> which validates event's id.
