+++ src/Netstr/Messaging/Events/Validators/EventPowValidator.cs
@@ -2,7 +2,7 @@
 using Netstr.Messaging.Models;
 using Netstr.Options;
 
-namespace Netstr.Messaging.MessageHandlers.Events
+namespace Netstr.Messaging.Events.Validators
 {
     /// <summary>
     /// <see cref="IEventValidator"/> which validates event's proof of work against congiured limits.
