+++ test/Netstr.Tests/Events/EventHandlersTests.cs
@@ -7,8 +7,8 @@ using Microsoft.Extensions.Options;
 using Moq;
 using Netstr.Data;
 using Netstr.Messaging;
-using Netstr.Messaging.EventHandlers;
-using Netstr.Messaging.EventHandlers.Replaceable;
+using Netstr.Messaging.Events.Handlers;
+using Netstr.Messaging.Events.Handlers.Replaceable;
 using Netstr.Messaging.Models;
 using Netstr.Messaging.WebSockets;
 using Netstr.Options;
